requirements.txt
Django==2.1.5
django-ckeditor==5.6.1
django-contrib-comments==1.9.0
django-js-asset==1.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
olefile==0.46
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
six==1.12.0
